chore: reduce controller flow to a mixin
