+++ renku/service/controllers/datasets_files_list.py
@@ -18,12 +18,13 @@
 """Renku service datasets files controller."""
 from renku.core.commands.dataset import list_files
 from renku.core.utils.contexts import chdir
+from renku.service.controllers.mixins import ReadOperationMixin
 from renku.service.controllers.remote_project import RemoteProject
 from renku.service.serializers.datasets import DatasetFilesListRequest, DatasetFilesListResponseRPC
 from renku.service.views import result_response
 
 
-class DatasetsFilesListCtrl:
+class DatasetsFilesListCtrl(ReadOperationMixin):
     """Controller for datasets files list endpoint."""
 
     REQUEST_SERIALIZER = DatasetFilesListRequest()
@@ -32,29 +33,16 @@ class DatasetsFilesListCtrl:
     def __init__(self, cache, user_data, request_data):
         """Construct a datasets files list controller."""
         self.ctx = DatasetsFilesListCtrl.REQUEST_SERIALIZER.load(request_data)
-        self.user = cache.ensure_user(user_data)
+        super(DatasetsFilesListCtrl, self).__init__(cache, user_data, request_data)
 
-        self.cache = cache
-        self.user_data = user_data
-        self.request_data = request_data
+    @property
+    def context(self):
+        """Controller operation context."""
+        return self.ctx
 
     def renku_op(self):
         """Renku operation for the controller."""
         return list_files(datasets=[self.ctx["name"]])
-
-    def local(self):
-        """Execute renku operation against service cache."""
-        project = self.cache.get_project(self.user, self.ctx["project_id"])
-
-        with chdir(project.abs_path):
-            return self.renku_op()
-
-    def remote(self):
-        """Execute renku operation against remote project."""
-        project = RemoteProject(self.user_data, self.request_data)
-
-        with project.remote():
-            return self.renku_op()
 
     def to_response(self):
         """Execute controller flow and serialize to service response."""
